requirements.txt
Flask==2.3.3
requests==2.31.0
pytest==7.4.2
pytest-html==4.0.2
python-dotenv==1.0.0
coverage==7.2.7